fix:#62 내역 상세 및 수정 프래그먼트 액티비티로 변경 + 버그 수정
