+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,8 @@
         <activity android:name="com.nexters.zzallang.harusal2.ui.budget.register.StartDayDefaultRegisterActivity"/>
         <activity android:name="com.nexters.zzallang.harusal2.ui.budget.register.StartDayClickRegisterActivity"/>
         <activity android:name=".ui.statement.register.AddStatementActivity"/>
-        <activity android:name=".ui.statement.edit.StatementActivity" />
+        <activity android:name=".ui.statement.edit.StatementDetailActivity" />
+        <activity android:name=".ui.statement.edit.StatementEditActivity" />
         <activity android:name=".ui.main.MainActivity" android:launchMode="singleTop"/>
         <activity android:name=".ui.budget.edit.BudgetEditActivity"/>
         <activity android:name=".ui.onBoarding.OnBoardingActivity" />
