+++ app/src/main/java/com/nexters/zzallang/harusal2/ui/statement/edit/StatementDetailViewModel.kt
@@ -54,4 +54,8 @@ class StatementDetailViewModel(private val statementUseCase: StatementUseCase): 
     fun setMemo(){
         _statementMemo.postValue(statement.content)
     }
+
+    suspend fun deleteStatement(){
+        statementUseCase.deleteStatement(getId())
+    }
 }
