--- app/src/main/java/com/nexters/zzallang/harusal2/ui/statement/edit/StatementEditActivity.java
@@ -1,4 +1,0 @@
-package com.nexters.zzallang.harusal2.ui.statement.edit;
-
-public class StatementEditActivity {
-}
