+++ app/src/main/res/layout/activity_edit_statement.xml
@@ -10,7 +10,8 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:background="@color/colorWhite">
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/header_statement"
             android:layout_width="match_parent"
