+++ app/src/main/res/layout/activity_statement_detail.xml
@@ -10,7 +10,8 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:background="@color/colorWhite">
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/header_statement"
             android:layout_width="match_parent"
@@ -84,17 +85,18 @@
 
             <TextView
                 android:id="@+id/tv_statement_type"
-                android:layout_width="48dp"
-                android:layout_height="22dp"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
                 app:layout_constraintLeft_toRightOf="@id/tv_statement_amount_title"
                 app:layout_constraintTop_toBottomOf="@id/tv_statement_date"
-                android:layout_marginTop="40dp"
+                android:layout_marginTop="42dp"
                 android:layout_marginLeft="8dp"
+                android:paddingLeft="6dp"
+                android:paddingRight="6dp"
                 android:text="@{vm.statementType}"
                 android:textSize="12sp"
                 android:textColor="@color/colorWhite"
-                android:textStyle="normal"
-                android:gravity="center">
+                android:textStyle="normal">
 
             </TextView>
