--- app/src/main/res/layout/fragment_statement_detail.xml
@@ -1,121 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-        <variable
-            name="vm"
-            type="com.nexters.zzallang.harusal2.ui.statement.edit.StatementDetailViewModel" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toTopOf="@id/btn_statement_detail_edit"
-            android:paddingLeft="20dp"
-            android:paddingRight="20dp">
-
-            <TextView
-                android:id="@+id/tv_statement_date"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintLeft_toLeftOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                android:layout_marginTop="32dp"
-                android:text="@{vm.statementDate}"
-                android:textSize="28sp"
-                android:textColor="@color/colorDarkBlack"
-                android:textStyle="normal">
-
-            </TextView>
-
-            <TextView
-                android:id="@+id/tv_statement_amount_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintLeft_toLeftOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/tv_statement_date"
-                android:layout_marginTop="40dp"
-                android:text="@string/amount_title"
-                android:textSize="16sp"
-                android:textStyle="normal"
-                android:textColor="@color/colorGray">
-
-            </TextView>
-            
-            <TextView
-                android:id="@+id/tv_statement_type"
-                android:layout_width="48dp"
-                android:layout_height="22dp"
-                app:layout_constraintLeft_toRightOf="@id/tv_statement_amount_title"
-                app:layout_constraintTop_toBottomOf="@id/tv_statement_date"
-                android:layout_marginTop="40dp"
-                android:layout_marginLeft="8dp"
-                android:text="@{vm.statementType}"
-                android:textSize="12sp"
-                android:textColor="@color/colorWhite"
-                android:textStyle="normal"
-                android:gravity="center">
-
-            </TextView>
-            
-            <TextView
-                android:id="@+id/tv_statement_amount"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintLeft_toLeftOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/tv_statement_amount_title"
-                android:layout_marginTop="10dp"
-                android:text="@{vm.statementAmount}"
-                android:textColor="@color/colorDarkBlack"
-                android:textSize="28sp"
-                android:textStyle="bold"/>
-
-            <TextView
-                android:id="@+id/tv_statement_memo_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintLeft_toLeftOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/tv_statement_amount"
-                android:layout_marginTop="40dp"
-                android:text="@string/memo_title"
-                android:textStyle="normal"
-                android:textSize="16sp"
-                android:textColor="@color/colorGray"/>
-
-            <TextView
-                android:id="@+id/tv_statement_memo"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintLeft_toLeftOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/tv_statement_memo_title"
-                android:layout_marginTop="12dp"
-                android:text="@{vm.statementMemo}"
-                android:textColor="@color/colorDarkBlack"
-                android:textSize="28sp"
-                android:textStyle="bold"/>
-            
-        </androidx.constraintlayout.widget.ConstraintLayout>
-        
-        <Button
-            android:id="@+id/btn_statement_detail_edit"
-            android:layout_width="match_parent"
-            android:layout_height="64dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:background="@color/colorBtnBlack"
-            android:text="@string/btn_text_edit"
-            android:textColor="@android:color/white"
-            android:textStyle="bold"
-            android:textSize="16sp"
-            >
-
-        </Button>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</layout>
